+++ pkgs/google_generative_ai/CHANGELOG.md
@@ -3,6 +3,7 @@
 - Update the package version that is sent with the HTTP client name.
 - Throw more actionable error objects than `FormatException` for errors. Errors
   were previously only correctly parsed in `generateContent` calls.
+- Add support for tuned models.
 
 ## 0.2.2